+++ tests/core/tools.js
@@ -58,6 +58,18 @@ bender.test(
 			assert.isFalse( CKEDITOR.tools.isArray( window.x ) );
 		},
 
+		'test getIndex - not found': function() {
+			assert.areSame( -1, CKEDITOR.tools.getIndex( [ 1, 2, 3 ], function( el ) { return el == 4; } ) );
+		},
+
+		'test getIndex - found first': function() {
+			assert.areSame( 2, CKEDITOR.tools.getIndex( [ 0, 1, 2, 2, 2, 3, 2, 2 ], function( el ) { return el == 2; } ) );
+		},
+
+		'test getIndex - found on last position': function() {
+			assert.areSame( 2, CKEDITOR.tools.getIndex( [ 0, 1, 2 ], function( el ) { return el == 2; } ) );
+		},
+
 		test_htmlEncode1: function() {
 			assert.areSame( '&lt;b&gt;Test&lt;/b&gt;', CKEDITOR.tools.htmlEncode( '<b>Test</b>' ) );
 		},