+++ tests/plugins/widget/widgetsrepoapi.js
@@ -157,6 +157,49 @@
 			assert.isTrue( true );
 		},
 
+		'test widgets.add - upcast priorities': function() {
+			var editor = this.editor,
+				bot = this.editorBot,
+				upcastCalled = '';
+
+			editor.widgets.add( 'prioritiesTest1', {
+				upcast: function() {
+					upcastCalled += '1';
+				}
+			} );
+
+			editor.widgets.add( 'prioritiesTest2', {
+				upcastPriority: 5,
+				upcast: function() {
+					upcastCalled += '2';
+				}
+			} );
+
+			editor.widgets.add( 'prioritiesTest3', {
+				upcastPriority: 5,
+				upcast: function() {
+					upcastCalled += '3';
+				}
+			} );
+
+			editor.widgets.add( 'prioritiesTest4', {
+				upcastPriority: 15,
+				upcast: function() {
+					upcastCalled += '4';
+				}
+			} );
+
+			editor.widgets.add( 'prioritiesTest5', {
+				upcast: function() {
+					upcastCalled += '5';
+				}
+			} );
+
+			bot.setData( '<p>foo</p>', function() {
+				assert.areSame( '23154', upcastCalled );
+			} );
+		},
+
 		'test widgets.wrapElement - inline': function() {
 			var editor = this.editor;